Draft prelude/foreword for Example 2
Signed-off-by: Kingdon Barrett <kingdon@weave.works>

more work on Example 2 README

We've been through the wasmer-ruby examples, and found out that we've
probably been thinking of Wasm all wrong.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

PLAN.txt

Reminder that we're doing these examples in service of building towards
a goal, once we know enough we can go back and complete the stats app!

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

and Gemfile for stats

(and if we get that graph going up and to the right as promised, we can
rebuild it in the form of a Kubernetes operator, a real crowd pleaser!)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

rust and Makefile

rustup, (you also need wabt to get past this point)
https://github.com/WebAssembly/wabt

try: make, make clean, make distclean

rustc is provided by rustup, it should be in your cargo path
wasm-strip and wasm-opt are Wasm binary tools from the WABT:
WebAssembly Binary Toolkit

(Myself, I had to compile and install that manually!)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

we'd like to add this code in

but I don't know about Cargo or crates yet (how did I get this far)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

distclean

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

make test

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

wrong but with more documentation

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

implement a Wasi with rust in Ruby

Not the hero we needed but the one we deserve!

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

let's try a wast

This way at least we can be sure we're invoking a wasm correctly, if it
exports a function with no params that returns i32 in the way expected!

This main.rs is wrong but test.rb is almost right

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

wast again - this time passing pointer to a string

I was so wrong about where this was going

I thought I'd be receiving an http client in the mail, turns out we're
going to be passing pointers around and it's not like that at all...

The `test.rb` is updated so there is no longer any expectation that we
take our http client as a parameter. We'll actually be exporting ours,
so the Rust/Wasm HTML parsing library that we've built can receive the
HTML as a pointer in memory. (That way we don't need to copy ... ?)

This wast at least works, we receive the "Hello World" string and the
Ruby program accepts it unquestioningly as a number to return inside of
the Ruby hash expressed as a to_string output:

{:time=>2023-04-24 19:41:40.018884 -0400, :count=>"Hello, World!"}

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

separate out the test scaffold from my wasm

We can isolate the parts that call out into Wasm-land from the parts
that our test scaffold took unto itself to handle on our behalf.

In a future version, we'll probably make my_wasmer a Spin module, and
it can be self-contained together with an HTTP client, so there's no
integration with Ruby required! It'll just run as a rust wasm module.
We call it with the name of the repo you want; it reads back a number!

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

attempt to use a workspace

https://stackoverflow.com/a/50402684/661659

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

make it build

now our library has multiple crate_type (lib, cdylib)
This means it can be linked to by other Rust (main.rs) and by consumers
who might be downstream targeting including us as a Wasm compiled module

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

no dashism

put the dash back, it's ok to call this lib-stat and refer to it as
lib_stat in contexts where dashes are not allowed (we're learning rs
syntax!)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

this should work

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

add pry

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

give up on bindgen for now

we can make the wasi program work in ruby wasmer, but not while we are
exporting bindings for Wasm consumers (maybe this can be resolved tho!)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

Merge Example 2 as Draft

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

tidy Makefile

Signed-off-by: Kingdon Barrett <kingdon@weave.works>