wast again - this time passing pointer to a string
I was so wrong about where this was going

I thought I'd be receiving an http client in the mail, turns out we're
going to be passing pointers around and it's not like that at all...

The `test.rb` is updated so there is no longer any expectation that we
take our http client as a parameter. We'll actually be exporting ours,
so the Rust/Wasm HTML parsing library that we've built can receive the
HTML as a pointer in memory. (That way we don't need to copy ... ?)

This wast at least works, we receive the "Hello World" string and the
Ruby program accepts it unquestioningly as a number to return inside of
the Ruby hash expressed as a to_string output:

{:time=>2023-04-24 19:41:40.018884 -0400, :count=>"Hello, World!"}

Signed-off-by: Kingdon Barrett <kingdon@weave.works>