separate out the test scaffold from my wasm
We can isolate the parts that call out into Wasm-land from the parts
that our test scaffold took unto itself to handle on our behalf.

In a future version, we'll probably make my_wasmer a Spin module, and
it can be self-contained together with an HTTP client, so there's no
integration with Ruby required! It'll just run as a rust wasm module.
We call it with the name of the repo you want; it reads back a number!

Signed-off-by: Kingdon Barrett <kingdon@weave.works>