add the Kubernetes YAML per the cli docs
$ be ruby cli
Commands:
  cli controller ORG  # Create a Project for the GitHub ORG and start reconciling it
  cli help [COMMAND]  # Describe available commands or one specific command

We implement this line:
Create a Project for the GitHub ORG and start reconciling it
---
if the Project hasn't been created yet, else just reconcile everything

The controller reconciles both Projects and Leaves

It also creates Leaf from each project, and each Leaf executes the Wasm
as well. The status on the Leaf gets updated then.

In the next iteration of this project iterator, we'll make sure that
Leaf and Project log their activity back to the database with Rails.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>