Completes in about 20s
fresh packages count: 0 (expecting 16) #######
fresh packages count: 7 (expecting 16) #######
final packages count: 16 (expecting 16) #######

The overall time spent is close to 20s now, but it will be much faster
on a Kubernetes control plane that is optimized and a better network.

The main delay is most likely due to the single-threadedness of Ruby
fibers. The concurrency model is a good experience, even if performance
can suffer!

Signed-off-by: Kingdon Barrett <kingdon@weave.works>