add placeholders for conditions
This should make it possible for children to record for their loving
parents what are the current results of their reconciling attempts.

It may make some new requirements on parents, for instance populating
the initial status with a reasonable structure with the required items
in order to validate the create/apply time resource when it is new (?)

No, I think you can just create a new resource without any Conditions
and that is "Normal" – it's up to the reconciler to create the Condition
items and reflect any changes with Condition updates to show an observed
generation for that condition.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>