it is faster to cross-build locally
it is much faster to build gems-base locally, then tag gem-cache and
make docker

Pushing layers takes a lot longer than anything else.

It ran bundle install and was finished in a second. It did not need Rust

The deploy image should be small and fast

It is taking over an hour to build gems as multiarch on GitHub, but very
thankfully we don't need to build the gems often. We can schedule it as
an overnight job that happens once a week, or more like twice a month!

Signed-off-by: Kingdon Barrett <kingdon@weave.works>